Add tests for JSON validator behavior wrt. numbers
Make sure that we test the JSON validator’s expected behavior with
respect to handling numbers. Currently, both JSON-typed numbers and
strings containing numbers are acceptable as numbers in the JSON
RPC-API, and we probably don’t want to break that.

This change will be accompanied by a companion change in the public
documentation in Zonemaster in order to make it clear that both numbers
and number-strings are acceptable for the RPCAPI.